Continue bloat cleanup for search config and recipes
