+++ src/config/search.rs
@@ -1,290 +1,12 @@
 use serde::{Deserialize, Serialize};
 use serde_json::json;
-use std::collections::BTreeSet;
-use std::thread::available_parallelism;
 
-use super::{env_string, parse_bool, parse_f32, parse_u64, parse_usize, set_string};
 use crate::error::{DbtNovaError, Result};
 
-const EXTENDED_META_DEFAULT_MAX_FIELDS: usize = 32;
-const EXTENDED_META_HARD_MAX_FIELDS: usize = 128;
-const EXTENDED_META_DEFAULT_MAX_VALUES_PER_FIELD: usize = 64;
-const EXTENDED_META_HARD_MAX_VALUES_PER_FIELD: usize = 1024;
-const EXTENDED_META_DEFAULT_MAX_BYTES_PER_VALUE: usize = 4096;
-const EXTENDED_META_HARD_MAX_BYTES_PER_VALUE: usize = 65_536;
+mod env;
+mod extended_meta;
 
-/// Supported value treatment for allowlisted non-Nova dbt metadata fields.
-#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ExtendedMetaFieldMode {
-    /// Index scalar metadata as exact/filterable keywords.
-    #[default]
-    Keyword,
-    /// Index scalar metadata as full text.
-    Text,
-    /// Index string arrays as repeated keyword values.
-    StringArray,
-    /// Index boolean metadata values.
-    Bool,
-}
-
-/// One allowlisted non-Nova dbt metadata path for extended-meta search.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
-#[serde(default)]
-pub struct ExtendedMetaFieldConfig {
-    /// Logical dbt metadata path, such as `meta.owner` or `columns.*.meta.semantic_group`.
-    pub path: String,
-    /// Stable search alias used for fielded search and optional summaries.
-    pub alias: String,
-    /// Value mode for this path.
-    pub mode: ExtendedMetaFieldMode,
-    /// Optional ranking boost applied to this field in search.
-    pub boost: f32,
-    /// Whether this field is included in `extended_meta_summary` payloads.
-    pub summary: bool,
-}
-
-impl Default for ExtendedMetaFieldConfig {
-    fn default() -> Self {
-        Self {
-            path: String::new(),
-            alias: String::new(),
-            mode: ExtendedMetaFieldMode::default(),
-            boost: 1.0,
-            summary: false,
-        }
-    }
-}
-
-/// Default-off extended metadata search configuration.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
-#[serde(default)]
-pub struct ExtendedMetaSearchConfig {
-    /// Explicit allowlist of non-Nova dbt metadata fields.
-    pub fields: Vec<ExtendedMetaFieldConfig>,
-    /// Maximum configured fields accepted.
-    pub max_fields: usize,
-    /// Maximum values indexed per configured field.
-    pub max_values_per_field: usize,
-    /// Maximum UTF-8 bytes retained per value.
-    pub max_bytes_per_value: usize,
-}
-
-impl Default for ExtendedMetaSearchConfig {
-    fn default() -> Self {
-        Self {
-            fields: Vec::new(),
-            max_fields: EXTENDED_META_DEFAULT_MAX_FIELDS,
-            max_values_per_field: EXTENDED_META_DEFAULT_MAX_VALUES_PER_FIELD,
-            max_bytes_per_value: EXTENDED_META_DEFAULT_MAX_BYTES_PER_VALUE,
-        }
-    }
-}
-
-impl ExtendedMetaSearchConfig {
-    /// Validate configured extended metadata paths and caps.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when the allowlist contains unsafe paths, duplicate
-    /// aliases, unsupported caps, or values that cannot be indexed safely.
-    pub fn validate(&self) -> Result<()> {
-        validate_bounded_usize(
-            "search.extended_meta.max_fields",
-            self.max_fields,
-            EXTENDED_META_HARD_MAX_FIELDS,
-        )?;
-        validate_bounded_usize(
-            "search.extended_meta.max_values_per_field",
-            self.max_values_per_field,
-            EXTENDED_META_HARD_MAX_VALUES_PER_FIELD,
-        )?;
-        validate_bounded_usize(
-            "search.extended_meta.max_bytes_per_value",
-            self.max_bytes_per_value,
-            EXTENDED_META_HARD_MAX_BYTES_PER_VALUE,
-        )?;
-
-        if self.fields.len() > self.max_fields {
-            return Err(DbtNovaError::InvalidParams(format!(
-                "search.extended_meta.fields configures {} fields but max_fields is {}",
-                self.fields.len(),
-                self.max_fields
-            )));
-        }
-
-        let mut aliases = BTreeSet::new();
-        let mut paths = BTreeSet::new();
-        for (index, field) in self.fields.iter().enumerate() {
-            validate_extended_meta_path(index, &field.path)?;
-            let alias = validate_extended_meta_alias(index, &field.alias)?;
-            if !aliases.insert(alias.to_string()) {
-                return Err(DbtNovaError::InvalidParams(format!(
-                    "search.extended_meta.fields[{index}].alias '{alias}' is configured more than once"
-                )));
-            }
-
-            let path = field.path.trim();
-            if !paths.insert(path.to_string()) {
-                return Err(DbtNovaError::InvalidParams(format!(
-                    "search.extended_meta.fields[{index}].path '{path}' is configured more than once"
-                )));
-            }
-
-            if !field.boost.is_finite() || field.boost < 0.0 {
-                return Err(DbtNovaError::InvalidParams(format!(
-                    "search.extended_meta.fields[{index}].boost must be a finite number greater than or equal to 0"
-                )));
-            }
-        }
-
-        Ok(())
-    }
-}
-
-fn validate_bounded_usize(name: &str, value: usize, hard_max: usize) -> Result<()> {
-    if value == 0 || value > hard_max {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "{name} must be between 1 and {hard_max}"
-        )));
-    }
-    Ok(())
-}
-
-fn validate_extended_meta_alias(index: usize, alias: &str) -> Result<&str> {
-    let alias = alias.trim();
-    if alias.is_empty() {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].alias is required"
-        )));
-    }
-    let mut chars = alias.chars();
-    let Some(first) = chars.next() else {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].alias is required"
-        )));
-    };
-    if !first.is_ascii_lowercase() {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].alias '{alias}' must start with a lowercase ASCII letter"
-        )));
-    }
-    if !chars.all(|ch| ch.is_ascii_lowercase() || ch.is_ascii_digit() || ch == '_') {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].alias '{alias}' must contain only lowercase ASCII letters, digits, and underscores"
-        )));
-    }
-
-    Ok(alias)
-}
-
-fn validate_extended_meta_path(index: usize, path: &str) -> Result<()> {
-    let path = path.trim();
-    if path.is_empty() {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].path is required"
-        )));
-    }
-
-    let segments = path.split('.').collect::<Vec<_>>();
-    if segments.iter().any(|segment| segment.trim().is_empty()) {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].path '{path}' contains an empty segment"
-        )));
-    }
-    for (segment_index, segment) in segments.iter().enumerate() {
-        let valid_segment = *segment == "*"
-            || segment
-                .chars()
-                .all(|ch| ch.is_ascii_alphanumeric() || ch == '_' || ch == '-');
-        if !valid_segment {
-            return Err(DbtNovaError::InvalidParams(format!(
-                "search.extended_meta.fields[{index}].path '{path}' must use dot-separated ASCII key names"
-            )));
-        }
-        if *segment == "*" && !(segment_index == 1 && segments.first() == Some(&"columns")) {
-            return Err(DbtNovaError::InvalidParams(format!(
-                "search.extended_meta.fields[{index}].path '{path}' may only use '*' in 'columns.*.meta.' paths"
-            )));
-        }
-        if let Some(sensitive) = contains_sensitive_segment(segment) {
-            return Err(DbtNovaError::InvalidParams(format!(
-                "search.extended_meta.fields[{index}].path '{path}' is not allowed because segment '{segment}' matches sensitive key '{sensitive}'"
-            )));
-        }
-    }
-
-    if !is_supported_extended_meta_path(&segments) {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].path '{path}' must start with 'meta.' or 'columns.*.meta.'"
-        )));
-    }
-    if is_nova_meta_path(&segments) {
-        return Err(DbtNovaError::InvalidParams(format!(
-            "search.extended_meta.fields[{index}].path '{path}' targets meta.nova, which is already indexed by Nova"
-        )));
-    }
-
-    Ok(())
-}
-
-fn is_supported_extended_meta_path(segments: &[&str]) -> bool {
-    if segments.len() >= 2 && segments[0] == "meta" {
-        return true;
-    }
-    segments.len() >= 4 && segments[0] == "columns" && segments[1] == "*" && segments[2] == "meta"
-}
-
-fn is_nova_meta_path(segments: &[&str]) -> bool {
-    (segments.len() >= 2 && segments[0] == "meta" && segments[1] == "nova")
-        || (segments.len() >= 4
-            && segments[0] == "columns"
-            && segments[1] == "*"
-            && segments[2] == "meta"
-            && segments[3] == "nova")
-}
-
-fn contains_sensitive_segment(segment: &str) -> Option<&'static str> {
-    let normalized = normalize_sensitive_segment(segment);
-    if normalized.contains("private_key") || normalized.contains("privatekey") {
-        return Some("private_key");
-    }
-    if normalized.contains("api_key") || normalized.contains("apikey") {
-        return Some("api_key");
-    }
-
-    for token in normalized.split('_') {
-        match token {
-            "token" => return Some("token"),
-            "secret" => return Some("secret"),
-            "password" => return Some("password"),
-            "credential" | "credentials" => return Some("credential"),
-            _ => {}
-        }
-    }
-    None
-}
-
-fn normalize_sensitive_segment(segment: &str) -> String {
-    let mut normalized = String::with_capacity(segment.len());
-    let mut previous_was_separator = true;
-    for ch in segment.chars() {
-        if ch.is_ascii_uppercase() {
-            if !previous_was_separator {
-                normalized.push('_');
-            }
-            normalized.push(ch.to_ascii_lowercase());
-            previous_was_separator = false;
-        } else if ch.is_ascii_alphanumeric() {
-            normalized.push(ch.to_ascii_lowercase());
-            previous_was_separator = false;
-        } else if !previous_was_separator {
-            normalized.push('_');
-            previous_was_separator = true;
-        }
-    }
-    normalized.trim_matches('_').to_string()
-}
+pub use extended_meta::{ExtendedMetaFieldConfig, ExtendedMetaFieldMode, ExtendedMetaSearchConfig};
 
 /// Weight multipliers for persona-aware ranking signals.
 #[derive(Debug, Clone, Copy, Deserialize, Serialize)]
@@ -836,7 +558,7 @@ impl Default for SearchConfig {
             vector_ann_max_candidates: 5000,
             vector_ann_min_candidates: 200,
             embedding_cache_dir: String::new(),
-            onnx_threads: default_onnx_threads(),
+            onnx_threads: env::default_onnx_threads(),
             embedding_batch_size: 128,
             sparse_embedding_batch_size: 16,
             enable_sparse_search: false,
@@ -855,7 +577,7 @@ impl SearchConfig {
     #[must_use]
     pub fn from_env() -> Self {
         let mut config = Self::default();
-        apply_search_env(&mut config);
+        env::apply_search_env(&mut config);
         config
     }
 
@@ -929,1036 +651,6 @@ impl SearchColdStartPolicy {
             _ => None,
         }
     }
-}
-
-fn default_onnx_threads() -> usize {
-    available_parallelism().map_or(1, |threads| threads.get().min(4))
-}
-
-fn apply_search_limits_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        max_page_size,
-        "DBT_NOVA_MAX_PAGE_SIZE",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        max_offset,
-        "DBT_NOVA_MAX_OFFSET",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        lineage_cache_size,
-        "DBT_NOVA_LINEAGE_CACHE_SIZE",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        column_lineage_precompute,
-        "DBT_NOVA_COLUMN_LINEAGE_PRECOMPUTE",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        default_limit,
-        "DBT_NOVA_DEFAULT_LIMIT",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        min_word_length,
-        "DBT_NOVA_MIN_WORD_LENGTH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        max_query_length,
-        "DBT_NOVA_MAX_QUERY_LENGTH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        max_path_pattern_length,
-        "DBT_NOVA_MAX_PATH_PATTERN_LENGTH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_search_index_env(config: &mut SearchConfig) {
-    set_string("DBT_NOVA_INDEX_DIR", &mut config.index_dir);
-    crate::env_config!(
-        config,
-        index_writer_heap_bytes,
-        "DBT_NOVA_INDEX_WRITER_HEAP_BYTES",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        max_sql_chunk_bytes,
-        "DBT_NOVA_MAX_SQL_CHUNK_BYTES",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        dedup_fetch_multiplier,
-        "DBT_NOVA_SEARCH_DEDUP_FETCH_MULTIPLIER",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_text_match_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        enable_ngram,
-        "DBT_NOVA_SEARCH_ENABLE_NGRAM",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        ngram_min,
-        "DBT_NOVA_SEARCH_NGRAM_MIN",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        ngram_max,
-        "DBT_NOVA_SEARCH_NGRAM_MAX",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        ngram_boost,
-        "DBT_NOVA_SEARCH_NGRAM_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        fuzzy_min_length,
-        "DBT_NOVA_FUZZY_MIN_LENGTH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        fuzzy_mid_length,
-        "DBT_NOVA_FUZZY_MID_LENGTH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        fuzzy_max_distance,
-        "DBT_NOVA_FUZZY_MAX_DISTANCE",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_highlight_format_env(config: &mut SearchConfig) {
-    if let Some(value) = env_string("DBT_NOVA_SEARCH_HIGHLIGHT_FORMAT") {
-        let normalized = value.trim().to_lowercase();
-        if matches!(normalized.as_str(), "html" | "text" | "plain") {
-            config.highlight_format = if normalized == "plain" {
-                "text".to_string()
-            } else {
-                normalized
-            };
-        }
-    }
-}
-
-fn apply_highlight_and_suggestion_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        highlight_max_chars,
-        "DBT_NOVA_SEARCH_HIGHLIGHT_MAX_CHARS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        highlight_max_fields,
-        "DBT_NOVA_SEARCH_HIGHLIGHT_MAX_FIELDS",
-        parse_usize
-    );
-    apply_highlight_format_env(config);
-    crate::env_config!(
-        config,
-        enable_suggestions,
-        "DBT_NOVA_SEARCH_ENABLE_SUGGESTIONS",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        suggestions_limit,
-        "DBT_NOVA_SEARCH_SUGGESTIONS_LIMIT",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_fusion_and_runtime_env(config: &mut SearchConfig) {
-    crate::env_config!(config, enable_rrf, "DBT_NOVA_SEARCH_ENABLE_RRF", parse_bool);
-    crate::env_config!(
-        config,
-        rrf_k,
-        "DBT_NOVA_SEARCH_RRF_K",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        rrf_overfetch,
-        "DBT_NOVA_SEARCH_RRF_OVERFETCH",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        search_circuit_failure_threshold,
-        "DBT_NOVA_SEARCH_CIRCUIT_FAILURE_THRESHOLD",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        search_circuit_open_seconds,
-        "DBT_NOVA_SEARCH_CIRCUIT_OPEN_SECONDS",
-        parse_u64,
-        |v: &u64| *v > 0
-    );
-    crate::env_config!(
-        config,
-        search_timeout_ms,
-        "DBT_NOVA_SEARCH_TIMEOUT_MS",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        search_max_concurrent,
-        "DBT_NOVA_SEARCH_MAX_CONCURRENT",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        search_max_queue,
-        "DBT_NOVA_SEARCH_MAX_QUEUE",
-        parse_usize
-    );
-}
-
-fn apply_vector_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        enable_vector_search,
-        "DBT_NOVA_SEARCH_ENABLE_VECTOR",
-        parse_bool
-    );
-    if let Some(value) = env_string("DBT_NOVA_SEARCH_COLD_START_POLICY")
-        && let Some(policy) = SearchColdStartPolicy::parse(&value)
-    {
-        config.cold_start_policy = policy;
-    }
-    crate::env_config!(
-        config,
-        vector_top_k,
-        "DBT_NOVA_SEARCH_VECTOR_TOP_K",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        vector_max_chars,
-        "DBT_NOVA_SEARCH_VECTOR_MAX_CHARS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        embeddings_max_decompressed_bytes,
-        "DBT_NOVA_EMBEDDINGS_MAX_DECOMPRESSED_BYTES",
-        parse_u64,
-        |v: &u64| *v > 0
-    );
-    crate::env_config!(
-        config,
-        enable_vector_ann,
-        "DBT_NOVA_SEARCH_ENABLE_VECTOR_ANN",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        enable_vector_quantization,
-        "DBT_NOVA_SEARCH_ENABLE_VECTOR_QUANTIZATION",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        vector_ann_bits,
-        "DBT_NOVA_SEARCH_VECTOR_ANN_BITS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        vector_ann_hamming,
-        "DBT_NOVA_SEARCH_VECTOR_ANN_HAMMING",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        vector_ann_max_candidates,
-        "DBT_NOVA_SEARCH_VECTOR_ANN_MAX_CANDIDATES",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        vector_ann_min_candidates,
-        "DBT_NOVA_SEARCH_VECTOR_ANN_MIN_CANDIDATES",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    set_string(
-        "DBT_NOVA_EMBEDDINGS_CACHE_DIR",
-        &mut config.embedding_cache_dir,
-    );
-    crate::env_config!(
-        config,
-        onnx_threads,
-        "DBT_NOVA_SEARCH_ONNX_THREADS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    set_string("DBT_NOVA_EMBEDDING_MODEL", &mut config.embedding_model);
-    crate::env_config!(
-        config,
-        embedding_batch_size,
-        "DBT_NOVA_SEARCH_EMBEDDING_BATCH_SIZE",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_sparse_and_reranker_env(config: &mut SearchConfig) {
-    let sparse_batch_override = std::env::var("DBT_NOVA_SEARCH_SPARSE_EMBEDDING_BATCH_SIZE")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .filter(|value| *value > 0);
-    let general_batch_override = std::env::var("DBT_NOVA_SEARCH_EMBEDDING_BATCH_SIZE")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .filter(|value| *value > 0);
-
-    crate::env_config!(
-        config,
-        enable_sparse_search,
-        "DBT_NOVA_SEARCH_ENABLE_SPARSE",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        sparse_top_k,
-        "DBT_NOVA_SEARCH_SPARSE_TOP_K",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        sparse_embedding_batch_size,
-        "DBT_NOVA_SEARCH_SPARSE_EMBEDDING_BATCH_SIZE",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    if sparse_batch_override.is_none()
-        && let Some(value) = general_batch_override
-    {
-        config.sparse_embedding_batch_size = value;
-    }
-    crate::env_config!(
-        config,
-        enable_reranker,
-        "DBT_NOVA_SEARCH_ENABLE_RERANKER",
-        parse_bool
-    );
-    set_string("DBT_NOVA_RERANKER_MODEL", &mut config.reranker_model);
-    crate::env_config!(
-        config,
-        rerank_top_n,
-        "DBT_NOVA_SEARCH_RERANK_TOP_N",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_persona_weights_env(var: &str, target: &mut PersonaWeights) {
-    if let Some(value) = env_string(var) {
-        target.apply_overrides(&value);
-    }
-}
-
-fn apply_persona_env(config: &mut SearchConfig) {
-    if let Some(value) = env_string("DBT_NOVA_SEARCH_DEFAULT_PERSONA") {
-        let trimmed = value.trim();
-        if !trimmed.is_empty() {
-            config.default_persona = Some(trimmed.to_string());
-        }
-    }
-    apply_persona_weights_env(
-        "DBT_NOVA_SEARCH_PERSONA_ANALYST_WEIGHTS",
-        &mut config.persona_weights.analyst,
-    );
-    apply_persona_weights_env(
-        "DBT_NOVA_SEARCH_PERSONA_ENGINEER_WEIGHTS",
-        &mut config.persona_weights.engineer,
-    );
-    apply_persona_weights_env(
-        "DBT_NOVA_SEARCH_PERSONA_GOVERNANCE_WEIGHTS",
-        &mut config.persona_weights.governance,
-    );
-    apply_persona_weights_env(
-        "DBT_NOVA_SEARCH_PERSONA_DEFAULT_WEIGHTS",
-        &mut config.persona_weights.default,
-    );
-}
-
-fn apply_field_boosts_env(config: &mut SearchConfig) {
-    apply_core_field_boosts_env(config);
-    apply_meta_field_boosts_env(config);
-}
-
-fn apply_core_field_boosts_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        field_boosts.alias,
-        "DBT_NOVA_ALIAS_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.name,
-        "DBT_NOVA_NAME_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.description,
-        "DBT_NOVA_DESCRIPTION_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.column,
-        "DBT_NOVA_COLUMN_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.tag,
-        "DBT_NOVA_TAG_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.path,
-        "DBT_NOVA_PATH_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.code,
-        "DBT_NOVA_CODE_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-}
-
-fn apply_meta_field_boosts_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        field_boosts.nova_synonyms,
-        "DBT_NOVA_META_SYNONYMS_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_domains,
-        "DBT_NOVA_META_DOMAINS_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_use_cases,
-        "DBT_NOVA_META_USE_CASES_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_measures,
-        "DBT_NOVA_META_MEASURES_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_metric,
-        "DBT_NOVA_META_METRIC_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_sensitivity,
-        "DBT_NOVA_META_SENSITIVITY_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_pii,
-        "DBT_NOVA_META_PII_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        field_boosts.nova_compliance,
-        "DBT_NOVA_META_COMPLIANCE_BOOST",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-}
-
-fn apply_semantic_scoring_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        staging_deboost_factor,
-        "DBT_NOVA_SEARCH_STAGING_DEBOOST_FACTOR",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_candidate_false_deboost_factor,
-        "DBT_NOVA_SEARCH_ANALYST_CANDIDATE_FALSE_DEBOOST_FACTOR",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        engineer_candidate_false_deboost_factor,
-        "DBT_NOVA_SEARCH_ENGINEER_CANDIDATE_FALSE_DEBOOST_FACTOR",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        governance_candidate_false_deboost_factor,
-        "DBT_NOVA_SEARCH_GOVERNANCE_CANDIDATE_FALSE_DEBOOST_FACTOR",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    apply_analyst_semantic_env(config);
-    apply_nova_semantic_env(config);
-}
-
-fn apply_analyst_semantic_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        analyst_semantic.metric_definition_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_METRIC_DEF_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.measure_definition_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_MEASURE_DEF_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.grain_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_GRAIN_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.time_field_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_TIME_FIELD_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.dimension_overlap_one_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_DIM_OVERLAP_ONE_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.dimension_overlap_two_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_DIM_OVERLAP_TWO_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.dimension_overlap_three_plus_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_DIM_OVERLAP_THREE_PLUS_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.missing_metric_or_measure_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_MISSING_METRIC_MEASURE_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.missing_grain_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_MISSING_GRAIN_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.min_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_MIN_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_semantic.max_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_MAX_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v > 0.0
-    );
-}
-
-fn apply_nova_semantic_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        nova_measure_match_multiplier,
-        "DBT_NOVA_SEARCH_MEASURE_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_metric_match_multiplier,
-        "DBT_NOVA_SEARCH_METRIC_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        analyst_nova_semantic_match_multiplier,
-        "DBT_NOVA_SEARCH_ANALYST_SEMANTIC_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        non_analyst_nova_semantic_match_multiplier,
-        "DBT_NOVA_SEARCH_NON_ANALYST_SEMANTIC_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_semantic_name_match_multiplier,
-        "DBT_NOVA_SEARCH_SEMANTIC_NAME_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_semantic_synonym_match_multiplier,
-        "DBT_NOVA_SEARCH_SEMANTIC_SYNONYM_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_semantic_definition_match_multiplier,
-        "DBT_NOVA_SEARCH_SEMANTIC_DEFINITION_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_semantic_canonical_match_multiplier,
-        "DBT_NOVA_SEARCH_SEMANTIC_CANONICAL_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_semantic_canonical_match_bonus,
-        "DBT_NOVA_SEARCH_SEMANTIC_CANONICAL_MATCH_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_synonym_match_multiplier,
-        "DBT_NOVA_SEARCH_SYNONYM_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_canonical_multiplier,
-        "DBT_NOVA_SEARCH_CANONICAL_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_canonical_match_multiplier,
-        "DBT_NOVA_SEARCH_CANONICAL_META_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        nova_canonical_match_bonus,
-        "DBT_NOVA_SEARCH_CANONICAL_META_MATCH_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        engineer_exact_match_multiplier,
-        "DBT_NOVA_SEARCH_ENGINEER_EXACT_MATCH_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-}
-
-fn apply_indicator_ranking_grouping_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        indicator_ranking.enable_parent_coherence,
-        "DBT_NOVA_SEARCH_INDICATOR_ENABLE_PARENT_COHERENCE",
-        parse_bool
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_group_max_groups,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_GROUP_MAX_GROUPS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_group_max_indicators,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_GROUP_MAX_INDICATORS",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_indicator_ranking_bonus_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        indicator_ranking.generic_label_bonus_one_token,
-        "DBT_NOVA_SEARCH_INDICATOR_GENERIC_LABEL_BONUS_ONE_TOKEN",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.generic_label_bonus_two_tokens,
-        "DBT_NOVA_SEARCH_INDICATOR_GENERIC_LABEL_BONUS_TWO_TOKENS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.generic_label_bonus_three_plus_tokens,
-        "DBT_NOVA_SEARCH_INDICATOR_GENERIC_LABEL_BONUS_THREE_PLUS_TOKENS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_synonym_bonus_one_token,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_SYNONYM_BONUS_ONE_TOKEN",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_synonym_bonus_two_tokens,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_SYNONYM_BONUS_TWO_TOKENS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_synonym_bonus_three_plus_tokens,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_SYNONYM_BONUS_THREE_PLUS_TOKENS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.semantic_label_precision_scale,
-        "DBT_NOVA_SEARCH_INDICATOR_SEMANTIC_LABEL_PRECISION_SCALE",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.semantic_label_precision_canonical_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_SEMANTIC_LABEL_PRECISION_CANONICAL_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-}
-
-fn apply_indicator_ranking_coherence_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_indicator_diversity_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_DIVERSITY_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_canonical_indicator_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_CANONICAL_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_strong_match_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_STRONG_MATCH_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_support_surface_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_SUPPORT_SURFACE_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_time_field_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_TIME_FIELD_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_dimension_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_DIMENSION_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.parent_coherence_max_bonus,
-        "DBT_NOVA_SEARCH_INDICATOR_PARENT_COHERENCE_MAX_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.search_parent_indicator_bonus_scale,
-        "DBT_NOVA_SEARCH_INDICATOR_SEARCH_PARENT_BONUS_SCALE",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.search_missing_indicator_parent_multiplier,
-        "DBT_NOVA_SEARCH_INDICATOR_SEARCH_MISSING_PARENT_MULTIPLIER",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.search_parent_indicator_top_k,
-        "DBT_NOVA_SEARCH_INDICATOR_SEARCH_PARENT_TOP_K",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.indicator_rrf_score_weight,
-        "DBT_NOVA_SEARCH_INDICATOR_RRF_SCORE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        indicator_ranking.indicator_reranker_score_weight,
-        "DBT_NOVA_SEARCH_INDICATOR_RERANKER_SCORE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-}
-
-fn apply_indicator_ranking_env(config: &mut SearchConfig) {
-    apply_indicator_ranking_grouping_env(config);
-    apply_indicator_ranking_bonus_env(config);
-    apply_indicator_ranking_coherence_env(config);
-}
-
-fn apply_metadata_support_env(config: &mut SearchConfig) {
-    crate::env_config!(
-        config,
-        metadata_support.parent_synonym_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_PARENT_SYNONYM_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.domain_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_DOMAIN_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.use_case_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_USE_CASE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.dimension_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_DIMENSION_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.column_name_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_COLUMN_NAME_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.column_role_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_COLUMN_ROLE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.semantic_type_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_SEMANTIC_TYPE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.example_value_weight,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_EXAMPLE_VALUE_WEIGHT",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.max_bonus,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_MAX_BONUS",
-        parse_f32,
-        |v: &f32| *v >= 0.0
-    );
-    crate::env_config!(
-        config,
-        metadata_support.max_values_per_field,
-        "DBT_NOVA_SEARCH_METADATA_SUPPORT_MAX_VALUES_PER_FIELD",
-        parse_usize,
-        |v: &usize| *v > 0
-    );
-}
-
-fn apply_extended_meta_env(config: &mut SearchConfig) {
-    if let Some(value) = env_string("DBT_NOVA_SEARCH_EXTENDED_META_FIELDS_JSON") {
-        match serde_json::from_str::<Vec<ExtendedMetaFieldConfig>>(&value) {
-            Ok(fields) => config.extended_meta.fields = fields,
-            Err(err) => {
-                config.env_errors.push(format!(
-                    "Invalid DBT_NOVA_SEARCH_EXTENDED_META_FIELDS_JSON JSON; expected a JSON array of extended metadata field objects with mode keyword|text|string_array|bool (error: {err})"
-                ));
-            }
-        }
-    }
-    crate::env_config!(
-        config,
-        extended_meta.max_fields,
-        "DBT_NOVA_SEARCH_EXTENDED_META_MAX_FIELDS",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        extended_meta.max_values_per_field,
-        "DBT_NOVA_SEARCH_EXTENDED_META_MAX_VALUES_PER_FIELD",
-        parse_usize
-    );
-    crate::env_config!(
-        config,
-        extended_meta.max_bytes_per_value,
-        "DBT_NOVA_SEARCH_EXTENDED_META_MAX_BYTES_PER_VALUE",
-        parse_usize
-    );
-}
-
-fn apply_search_env(config: &mut SearchConfig) {
-    apply_search_limits_env(config);
-    apply_search_index_env(config);
-    apply_text_match_env(config);
-    apply_highlight_and_suggestion_env(config);
-    apply_fusion_and_runtime_env(config);
-    apply_vector_env(config);
-    apply_sparse_and_reranker_env(config);
-    apply_persona_env(config);
-    apply_field_boosts_env(config);
-    apply_semantic_scoring_env(config);
-    apply_indicator_ranking_env(config);
-    apply_metadata_support_env(config);
-    apply_extended_meta_env(config);
 }
 
 #[cfg(test)]
